requirements.txt
pandas
openpyxl
pyarrow
scikit-learn
sentence-transformers
faiss-cpu
ollama
streamlit
